Moved helper functions to another file
- Moved all helper functions (non-Flask functions) to helper_functions.py
- Imported helper_functions to main.py as helpers
- Removed old authentication flow
